Speed up NamedMatrix updates between 40x and 200x (#2268)
* Factorize named matrix vote normalization option

The tests are also fixed, while keeping the same behavior as before.

Weirdly, update() does not normalize the values being set, whereas batch_update() does.
And _convert_to_numeric() keeps NaN values as NaN, whereas batch_update() converts them to 0.0 by default.
This is not very consistent, but I have kept the same behavior for backward compatibility.
Since not all Delphi tests are passing, I could not verify whether other parts of the pipeline depend on this behaviour.

* Speed up named matrix computation

Keep both behaviours in this commit, for comparison and to log a speed report.
Will remove it before pull-request.

* Add deep test and remove speed up comparison

This concludes the refactoring.

* Apply copilot spelling corrections